Drop tilt dependency
